Split Intermediate Timer and Catch Event tests
